requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
opensearch-py>=2.0.0
openpyxl>=3.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
anthropic>=0.18.0
